srcs/api-gateway-app/requirements.txt
flask==2.3.0
requests==2.31.0
pika==1.3.1
python-dotenv==1.0.0
